Benchmarks: migrate Total/UnitPrice/LineTotal from decimal to double
Phase 2 of benchmark-double-migration. Switches the money-shaped columns
on OrderSchema/OrderItemSchema and their EF/DTO mirrors from decimal to
double, and updates DatabaseSetup seed literals (10.0m/1.5m/5.0m/2.5m
→ 10.0/1.5/5.0/2.5) accordingly. The DapperOrderLagDto workaround class
is removed; the regular OrderLagDto now uses double/double? so Dapper
no longer needs a parallel type.

This removes the SQLite GetDecimal(string-parse) per-cell tax from every
reader on the benchmark hot path, so the inter-library comparison
measures library overhead rather than a driver implementation choice.

Quarry.Benchmarks does NOT compile in isolation after this commit (the
reader bodies still call GetDecimal/ExecuteScalarAsync<decimal> and
three benchmark files remain parked as `.cs.disabled`). Phase 3 restores
compilation. Quarry.Tests builds and the full test suite still passes.

+++ src/Quarry.Benchmarks/Infrastructure/DatabaseSetup.cs
@@ -61,7 +61,7 @@ public static class DatabaseSetup
             cmd.CommandText = "INSERT INTO orders (UserId, Total, Status, OrderDate) VALUES (@userId, @total, @status, @date)";
             cmd.Parameters.Clear();
             cmd.Parameters.AddWithValue("@userId", i);
-            cmd.Parameters.AddWithValue("@total", Math.Round(10.0m + (i * 1.5m), 2));
+            cmd.Parameters.AddWithValue("@total", Math.Round(10.0 + (i * 1.5), 2));
             cmd.Parameters.AddWithValue("@status", Statuses[i % Statuses.Length]);
             cmd.Parameters.AddWithValue("@date", DateTime.UtcNow.AddDays(-i).ToString("yyyy-MM-dd HH:mm:ss"));
             cmd.ExecuteNonQuery();
@@ -75,7 +75,7 @@ public static class DatabaseSetup
             for (int j = 1; j <= itemCount; j++)
             {
                 itemId++;
-                decimal unitPrice = Math.Round(5.0m + (itemId % 20) * 2.5m, 2);
+                double unitPrice = Math.Round(5.0 + (itemId % 20) * 2.5, 2);
                 int qty = 1 + (itemId % 5);
                 cmd.CommandText = "INSERT INTO order_items (OrderId, ProductName, Quantity, UnitPrice, LineTotal) VALUES (@orderId, @name, @qty, @price, @total)";
                 cmd.Parameters.Clear();
